Task/requirements.txt
django>=5.0,<7.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
channels>=4.0
channels-redis>=4.1
daphne>=4.0
redis>=5.0
Pillow>=10.0
